Remove wrapper as it loss pivot values with belongs to many
