+++ src/QueryBuilder.php
@@ -108,9 +108,7 @@ class QueryBuilder implements ArrayAccess
 
     public function __call($name, $arguments)
     {
-        $subject = new QueryBuilderPaginator($this->subject, $this->request);
-
-        $result = $this->forwardCallTo($subject, $name, $arguments);
+        $result = $this->forwardCallTo($this->subject, $name, $arguments);
 
         /*
          * If the forwarded method call is part of a chain we can return $this
@@ -129,11 +127,41 @@ class QueryBuilder implements ArrayAccess
         }
 
         if ($result instanceof LengthAwarePaginator || $result instanceof Paginator || $result instanceof CursorPaginator) {
+            $result->appends($this->getQueryBuilderParamsFromRequest());
             $this->addAppendsToResults(collect($result->items()));
         }
 
 
         return $result;
+    }
+
+    private function getQueryBuilderParamsFromRequest()
+    {
+        return array_filter(
+            array_merge(
+                $this->request->filters()->mapWithKeys(function ($value, $key) {
+                    return [config('query-builder.parameters.filter') . "[${key}]" => $value];
+                })->toArray(),
+                [
+                    config('query-builder.parameters.sort') => $this->request->sorts()->join(
+                        QueryBuilderRequest::getIncludesArrayValueDelimiter()
+                    )
+                ],
+                [
+                    config('query-builder.parameters.append') => $this->request->appends()->join(
+                        QueryBuilderRequest::getAppendsArrayValueDelimiter()
+                    )
+                ],
+                [
+                    config('query-builder.parameters.include') => $this->request->includes()->join(
+                        QueryBuilderRequest::getIncludesArrayValueDelimiter()
+                    )
+                ],
+                $this->request->fields()->mapWithKeys(function ($values, $key) {
+                    return [config('query-builder.parameters.fields') . "[${key}]" => implode(",", $values)];
+                })->toArray(),
+            )
+        );
     }
 
     public function clone()
